fix(error): 🐛  Resolve issue with duplicated re-export
