+++ packages/runtime/source/environment/environment.ts
@@ -1,7 +1,5 @@
+import { RuntimeError } from "@terminal-nerds/snippets-error/custom";
 import { z } from "zod";
-
-import { RuntimeError } from "../error/error.ts";
-
 /* prettier-ignore */
 export const RUNTIME_ENVIRONMENTS = [
 	"browser",
