--- packages/runtime/source/error/error.ts
@@ -1,1 +1,0 @@
-export { isRuntimeError, RUNTIME_ERROR_SCHEMA, RuntimeError } from "@terminal-nerds/snippets-error/custom";
