+++ packages/runtime/source/main.ts
@@ -1,6 +1,5 @@
 /* MODULES */
 export * from "./environment/environment.ts";
-export * from "./error/error.ts";
 export * from "./module/module.ts";
 export * from "./scope/scope.ts";
 export * from "./variable/variable.ts";
