+++ packages/runtime/source/variable/variable.ts
@@ -1,5 +1,6 @@
+import { RuntimeError } from "@terminal-nerds/snippets-error/custom";
+
 import { getRuntimeEnvironmentName, type RuntimeEnvironment } from "../environment/environment.ts";
-import { RuntimeError } from "../error/error.ts";
 
 const SUPPORTED_RUNTIMES = ["bun", "deno", "node"] as const;
 type SupportedRuntime = Extract<RuntimeEnvironment, (typeof SUPPORTED_RUNTIMES)[number]>;
